Move NFSv3 accounts to a component resource
Diagnostic settings are created as part of the component.
