+++ data_safe_haven/infrastructure/components/composite/__init__.py
@@ -9,6 +9,10 @@ from .microsoft_sql_database import (
     MicrosoftSQLDatabaseProps,
 )
 from .nfsv3_blob_container import NFSV3BlobContainerComponent, NFSV3BlobContainerProps
+from .nfsv3_storage_account import (
+    NFSV3StorageAccountComponent,
+    NFSV3StorageAccountProps,
+)
 from .postgresql_database import PostgresqlDatabaseComponent, PostgresqlDatabaseProps
 from .virtual_machine import LinuxVMComponentProps, VMComponent
 
@@ -23,6 +27,8 @@ __all__ = [
     "MicrosoftSQLDatabaseProps",
     "NFSV3BlobContainerComponent",
     "NFSV3BlobContainerProps",
+    "NFSV3StorageAccountComponent",
+    "NFSV3StorageAccountProps",
     "PostgresqlDatabaseComponent",
     "PostgresqlDatabaseProps",
     "VMComponent",
