+++ data_safe_haven/infrastructure/components/wrapped/__init__.py
@@ -1,7 +1,5 @@
 from .log_analytics_workspace import WrappedLogAnalyticsWorkspace
-from .nfsv3_storage_account import WrappedNFSV3StorageAccount
 
 __all__ = [
     "WrappedLogAnalyticsWorkspace",
-    "WrappedNFSV3StorageAccount",
 ]
